Allow Player to move

+++ minecraft-server/src/position.rs
@@ -36,9 +36,9 @@ pub struct BlockPositionInChunk {
 }
 
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 pub struct ChunkPosition {
